feat:rename components and functions to ui and logic
